+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Authentication tester dialog.
 
+### Changed
+- Promoted to Beta
 
 ## [0.6.0] - 2023-05-09
 ### Added